test/utils.ts was used elsewhere for non-tests, so renaming to mock-doc instead
