+++ packages/html/test/html-recipes.test.ts
@@ -1,6 +1,6 @@
 import { afterEach, beforeEach, describe, it } from "@std/testing/bdd";
 import { render, VNode } from "../src/index.ts";
-import { MockDoc } from "../src/utils.ts";
+import { MockDoc } from "../src/mock-doc.ts";
 import {
   type Cell,
   createBuilder,
